The issue of web phishing assaults has developed significantly as of  late and phishing is considered as one of the most unsafe web violations, which may cause gigantic what’s more, negative impact on online business. In a web phishing assault, the phisher makes a produced or phishing site to cheat web clients so as to acquire their touchy budgetary and individual data. A few traditional procedures for recognizing phishing site have been proposed to adapt to this issue. In any case, classification of phishing site is a difficult as large portion of these methods are not ready to make an exact choice powerfully concerning whether the new site is phishing or genuine. This paper reflects a strategy for phishing site recognition depends on IOT technology by using Raspberry Pi and  supervised machine learning classification. The experimental results demonstrated that the performance of the machine learning classifiers was improved by using the supervised features selection. Moreover, the machine learning classifiers with the supervised features selection outperformed the machine learning classifiers with other features selection methods.
The issue of web phishing assaults has developed significantly as of  late and phishing is considered as one of the most unsafe web violations, which may cause gigantic what’s more, negative impact on online business. In a web phishing assault, the phisher makes a produced or phishing site to cheat web clients so as to acquire their touchy budgetary and individual data. A few traditional procedures for recognizing phishing site have been proposed to adapt to this issue. In any case, classification of phishing site is a difficult as large portion of these methods are not ready to make an exact choice powerfully concerning whether the new site is phishing or genuine. This paper reflects a strategy for phishing site recognition depends on IOT technology by using Raspberry Pi and  supervised machine learning classification. The experimental results demonstrated that the performance of the machine learning classifiers was improved by using the supervised features selection. Moreover, the machine learning classifiers with the supervised features selection outperformed the machine learning classifiers with other features selection methods.